pyproject.toml
[project]
name = "contrataiiot"
version = "0.1.0"
description = "A Langraph + LangChain project using Ollama models to simulate a real-life application where you can translate all your contracts into simple words"
readme = "README.md"
requires-python = "==3.11"
dependencies = [
    "chromadb>=1.3.4",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.3",
    "ollama>=0.6.1",
    "pdfplumber>=0.11.8",
    "pymongo[srv]>=4.15.4",
    "python-dotenv>=1.2.1",
    "questionary>=2.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=3.0.0",
    "tenacity<8.4.0",
]
